requirements.txt
colorama
googlesearch-python
urllib3==1.26.15
six
